split out modes
