+++ src/modes/turtle.ts
@@ -1,0 +1,9 @@
+// Copyright (c) 2022 Dane Freeman.
+// Distributed under the terms of the Modified BSD License.
+
+/**
+ * Add additional mappings for Turtle.
+ */
+export async function installSPARQL(_CodeMirror: any) {
+  // presently a no-op
+}
